[UPDATE] send email with verification token is now a task out of api main thread
