+++ subroutines/templates/emails/users/account_verification.html
@@ -2,8 +2,8 @@
 
 <p>
   Before you start using <b>Sub Routines</b> we need you to do one last thing.
-  Please visit the following link <a href="">subroutines.com</a> and enter this
-  code
+  Please visit the following link <a href="{{ url }}">{{ url }}</a> and enter this
+  code:
 </p>
 
 <code>{{ token }}</code>
