--- subroutines/users/serializers/__init__.py
@@ -1,6 +1,0 @@
-from .users import (
-    UserSerializer,
-    UserSignUpSerializer,
-    UserLoginSerializer,
-    AccountVerificationSerializer,
-)
